refactor & remove test docs:)
