--- docs/projectA/README.md
@@ -1,10 +1,0 @@
----
-id: readme
-title: README
----
-
-## Docs
-Welcome to Project A, here are some available docs
-
-1. [Architecture](projectA/docs/architecture.md)
-2. [Contributors](projectA/docs/contributors.md)
