+++ lib/server/readMetadata.js
@@ -15,7 +15,7 @@ const chalk = require('chalk');
 const env = require('./env.js');
 const siteConfig = require(CWD + '/siteConfig.js');
 const versionFallback = require('./versionFallback.js');
-const escapeStringRegexp = require('escape-string-regexp');
+const utils = require('./utils.js');
 
 const SupportedHeaderFields = new Set([
   'id',
@@ -121,39 +121,10 @@ function extractMetadata(content) {
   return {metadata, rawContent: both.content};
 }
 
-// Return the subdirectory path of a doc file
-// Example: `docs/projectA/test.md` subDir is `projectA`
-function getSubDir(file) {
-  let subDir = path.dirname(
-    path.relative(path.join(CWD, '../', getDocsPath()), file)
-  );
-  // Convert Windows backslash paths to slash paths
-  subDir = subDir.replace('\\', '/');
-  return subDir !== '.' ? subDir : null;
-}
-
-// process the metadata for a document found in the docs folder
-function processMetadata(file) {
+// process the metadata for a document found in either 'docs' or 'translated_docs'
+function processMetadata(file, refDir) {
   const result = extractMetadata(fs.readFileSync(file, 'utf8'));
-
-  let regexSubFolder = new RegExp(
-    '/' + escapeStringRegexp(getDocsPath()) + '/(.*)/.*/'
-  );
-
-  const match = regexSubFolder.exec(file);
-  let language = 'en';
-
-  // Avoid misinterpreting subdirectory as language
-  if (match && env.translation.enabled) {
-    const enabledLanguages = env.translation
-      .enabledLanguages()
-      .map(language => language.tag);
-
-    if (enabledLanguages.indexOf(match[1]) !== -1) {
-      language = match[1];
-    }
-  }
-
+  const language = utils.getLanguage(file, refDir) || 'en';
   const metadata = {};
   for (const fieldName of Object.keys(result.metadata)) {
     if (SupportedHeaderFields.has(fieldName)) {
@@ -172,9 +143,11 @@ function processMetadata(file) {
     throw new Error('Document id cannot include "/".');
   }
 
-  // If a doc file is located in a subdirectory, prepend the subdir to it's ID
-  // Example: `docs/projectA/test.md` with 'test' ID is changed to `projectA/test`
-  const subDir = getSubDir(file);
+  // If a file is located in a subdirectory, prepend the subdir to it's ID
+  // Example:
+  //  (file: 'docusaurus/docs/projectA/test.md', ID 'test', refDir: 'docs')
+  //  returns 'projectA/test'
+  const subDir = utils.getSubDir(file, refDir);
   if (subDir) {
     metadata.id = subDir + '/' + metadata.id;
   }
@@ -250,7 +223,7 @@ function generateMetadataDocs() {
     const extension = path.extname(file);
 
     if (extension === '.md' || extension === '.markdown') {
-      const res = processMetadata(file);
+      const res = processMetadata(file, path.join(CWD, '../', getDocsPath()));
 
       if (!res) {
         return;
@@ -289,23 +262,17 @@ function generateMetadataDocs() {
   });
 
   // metadata for non-english docs
-  const regexSubFolder = /translated_docs\/(.*?)\/.*/;
+  const translatedDir = path.join(CWD, 'translated_docs');
   files = glob.sync(CWD + '/translated_docs/**');
   files.forEach(file => {
-    let language = 'en';
-    const match = regexSubFolder.exec(file);
-    if (match) {
-      language = match[1];
-    }
-
-    if (enabledLanguages.indexOf(language) === -1) {
+    if (!utils.getLanguage(file, translatedDir)) {
       return;
     }
 
     const extension = path.extname(file);
 
     if (extension === '.md' || extension === '.markdown') {
-      const res = processMetadata(file);
+      const res = processMetadata(file, translatedDir);
       if (!res) {
         return;
       }
@@ -445,7 +412,6 @@ function generateMetadataBlog() {
 
 module.exports = {
   getDocsPath,
-  getSubDir,
   readSidebar,
   extractMetadata,
   processMetadata,
