+++ lib/server/versionFallback.js
@@ -12,6 +12,7 @@ const path = require('path');
 const assert = require('assert');
 
 const env = require('./env.js');
+const utils = require('./utils.js');
 const siteConfig = require(CWD + '/siteConfig.js');
 
 const ENABLE_TRANSLATION = fs.existsSync(CWD + '/languages.js');
@@ -190,17 +191,6 @@ function diffLatestDoc(file, id) {
   );
 }
 
-// Return the subdirectory path of a versioned_doc file
-// Example: `versioned_docs/version-1.0.11/projectA/test.md` subDir is `projectA`
-function getSubDir(file, version) {
-  let subDir = path.dirname(
-    path.relative(path.join(CWD, 'versioned_docs', `version-${version}`), file)
-  );
-  // Convert Windows backslash paths to slash paths
-  subDir = subDir.replace('\\', '/');
-  return subDir !== '.' ? subDir : null;
-}
-
 // return metadata for a versioned file given the file, its version (requested),
 // the version of the file to be used, and its language
 function processVersionMetadata(file, version, useVersion, language) {
@@ -209,7 +199,10 @@ function processVersionMetadata(file, version, useVersion, language) {
   // Add subdirectory information to versioned_doc metadata
   // Example: `versioned_docs/version-1.1.6/projectA/readme.md` file with id `version-1.1.6-readme`
   // and original_id `readme` will have metadata id of `version-1.1.6-projectA/readme` and original_id `projectA/readme`
-  const subDir = getSubDir(file, useVersion);
+  const subDir = utils.getSubDir(
+    file,
+    path.join(CWD, 'versioned_docs', `version-${useVersion}`)
+  );
   if (subDir) {
     metadata.original_id = subDir + '/' + metadata.original_id;
     metadata.id = metadata.id.replace(
