+++ lib/write-translations.js
@@ -52,7 +52,8 @@ function execute() {
     if (extension === '.md' || extension === '.markdown') {
       let res;
       try {
-        res = readMetadata.processMetadata(file);
+        const docsDir = path.join(CWD, '../', readMetadata.getDocsPath());
+        res = readMetadata.processMetadata(file, docsDir);
       } catch (e) {
         console.error(e);
         process.exit(1);
